refactor(MenuGroupService) - MenuGroupService.findById에orElseThrow 변경 (IllegalArgumentException -> MenuGroupNotFoundException) - test 변경
